refactor(backend): extract task retry & failure to services/tasks/retry.ts (U1)
Part of the tasks.ts mega-file split (CQ-H1). Moves the retry/failure
machinery -- handleTaskFailure, reassignStaleTasks, MAX_RETRIES, plus
the private helpers (readKeyspaceProgress, readWorkRangeField,
terminalFailStaleTask, StaleTaskRow) -- into a sibling submodule.

Also moves jsonSafeBigint + SAFE_NUMBER_THRESHOLD to a tiny
tasks/_internals.ts so the parent and retry both import from the same
definition instead of cycling through the barrel. The underscore prefix
keeps it out of the public surface.

services/tasks.ts shrinks from 1286 -> 884 LOC; remaining overshoot
will be addressed by U2 (per-agent projection, ~81 LOC out) and U3
(zap endpoint, ~57 LOC out) -- target parent LOC ~746.

Re-exports from tasks.ts keep heartbeat.ts's lazy import and the seven
mock.module test sites binding identically. No behavior change.

Validation: just check green, 564 backend tests pass, isolated-phase
tasks.test.ts (30/30) and agent-heartbeat integration (17/17) pass.
