+++ packages/backend/src/services/tasks.ts
@@ -20,23 +20,12 @@ import { updateCampaignProgress } from './campaigns.js'
 import { pickChunkSize } from './chunk-sizing.js'
 import { emitCrackResult, emitTaskUpdate } from './events.js'
 import { calculateAttackKeyspace } from './keyspace.js'
+import { jsonSafeBigint } from './tasks/_internals.js'
 
 // ─── Task Generation ────────────────────────────────────────────────
 
 // Below this threshold, workRange fields can be stored as JS Number safely
 // without losing precision (Number.MAX_SAFE_INTEGER = 2^53 - 1).
-const SAFE_NUMBER_THRESHOLD = BigInt(Number.MAX_SAFE_INTEGER)
-
-/**
- * Pick the JSON representation of a bigint value: a JS Number when the value
- * fits in JS-safe integer range, otherwise a decimal string. Keeps existing
- * `tasks.workRange` consumers (which read fields as numbers) working for
- * realistic-sized attacks while preserving precision for mask keyspaces that
- * overflow Number.MAX_SAFE_INTEGER.
- */
-function jsonSafeBigint(value: bigint): number | string {
-  return value <= SAFE_NUMBER_THRESHOLD ? Number(value) : value.toString()
-}
 
 /**
  * Derives required capabilities from an attack's configuration.
@@ -672,406 +661,6 @@ export async function updateTaskProgress(
   return { task: updated }
 }
 
-// ─── Task Retry & Failure Handling ──────────────────────────────────
-
-/**
- * Maximum reassignment count before a task is permanently failed. A task
- * with `retry_count >= MAX_RETRIES` reaching either failure path
- * (`handleTaskFailure` or the stale-task sweep) is marked terminal.
- * Exported so callers and tests can reason about the same bound.
- */
-export const MAX_RETRIES = 3
-
-export async function handleTaskFailure(taskId: number, agentId: number, reason: string) {
-  const [task] = await db
-    .select()
-    .from(tasks)
-    .where(and(eq(tasks.id, taskId), eq(tasks.agentId, agentId)))
-    .limit(1)
-  if (!task) {
-    return { error: 'Task not found or not assigned to this agent' }
-  }
-
-  const resultStats = (task.resultStats as Record<string, unknown>) ?? {}
-  const { retryCount } = task
-
-  // Derive projectId from the campaign for event emission
-  const [campaign] = await db
-    .select({ projectId: campaigns.projectId })
-    .from(campaigns)
-    .where(eq(campaigns.id, task.campaignId))
-    .limit(1)
-
-  if (retryCount < MAX_RETRIES) {
-    // Retry: reset task to pending with incremented retry count
-    const [updated] = await db
-      .update(tasks)
-      .set({
-        status: 'pending',
-        agentId: null,
-        assignedAt: null,
-        startedAt: null,
-        failureReason: reason,
-        retryCount: retryCount + 1,
-        resultStats: { ...resultStats, lastFailure: reason },
-        updatedAt: new Date(),
-      })
-      .where(and(eq(tasks.id, taskId), eq(tasks.agentId, agentId)))
-      .returning()
-
-    if (updated && campaign) {
-      // Surface the agent that was just freed so listeners can refresh that
-      // agent's caches; the row itself no longer holds agentId after retry.
-      emitTaskUpdate(campaign.projectId, taskId, 'pending', {
-        agentId,
-        campaignId: task.campaignId,
-      })
-    }
-
-    return { task: updated, retried: true }
-  }
-
-  // Max retries exceeded — mark as failed permanently. Use the stable
-  // terminal code `max_retries_exceeded` so this row is distinguishable
-  // from a one-shot failure with the same underlying cause (the sweep's
-  // terminal branches use the same code). The agent-reported reason that
-  // tipped the task over the budget is preserved in resultStats.lastFailure
-  // for debugging.
-  //
-  // Guard the UPDATE on agentId + status so a concurrent sweep that
-  // already reassigned this row cannot cause us to mark a now-unowned
-  // task as failed; only emit events when a row was actually updated.
-  const [updated] = await db
-    .update(tasks)
-    .set({
-      status: 'failed',
-      failureReason: 'max_retries_exceeded',
-      completedAt: new Date(),
-      resultStats: { ...resultStats, lastFailure: reason },
-      updatedAt: new Date(),
-    })
-    .where(
-      and(
-        eq(tasks.id, taskId),
-        eq(tasks.agentId, agentId),
-        sql`${tasks.status} IN ('assigned', 'running')`
-      )
-    )
-    .returning()
-
-  if (updated && campaign) {
-    emitTaskUpdate(campaign.projectId, taskId, 'failed', {
-      agentId,
-      campaignId: task.campaignId,
-    })
-    // Permanent fail changes the active-task count for the campaign; refresh
-    // the aggregate so dashboards do not lag a sweep cycle. The sweep's
-    // terminal-fail branches already do this — keep the two failure paths
-    // symmetric so either subsystem keeps the aggregate honest.
-    await updateCampaignProgress(task.campaignId)
-  }
-
-  return { task: updated, retried: false }
-}
-
-/**
- * Read the keyspace-progress value from a task's `progress` jsonb. Accepts
- * either a number or a numeric string (older agents may emit either).
- * Returns 0 for missing / unparseable values so the caller can treat the
- * task as "fresh" rather than fail noisily.
- */
-function readKeyspaceProgress(progress: unknown): bigint {
-  if (progress === null || typeof progress !== 'object') return 0n
-  const raw = (progress as Record<string, unknown>)['keyspaceProgress']
-  if (typeof raw === 'number' && Number.isFinite(raw)) return BigInt(Math.floor(raw))
-  if (typeof raw === 'string') {
-    try {
-      return BigInt(raw)
-    } catch {
-      return 0n
-    }
-  }
-  return 0n
-}
-
-/**
- * Read a numeric field from a task's `work_range` jsonb. Accepts either a
- * JS Number (used for in-safe-range chunks) or a decimal string (used for
- * mask-attack chunks beyond Number.MAX_SAFE_INTEGER).
- */
-function readWorkRangeField(workRange: unknown, key: string): bigint {
-  if (workRange === null || typeof workRange !== 'object') return 0n
-  const raw = (workRange as Record<string, unknown>)[key]
-  if (typeof raw === 'number' && Number.isFinite(raw)) return BigInt(Math.floor(raw))
-  if (typeof raw === 'string') {
-    try {
-      return BigInt(raw)
-    } catch {
-      return 0n
-    }
-  }
-  return 0n
-}
-
-/**
- * Stale-task descriptor carried through the sweep loop. Includes the
- * agent's current ownership so per-task UPDATEs can be guarded against a
- * concurrent sweep already reaping the same row.
- */
-type StaleTaskRow = {
-  taskId: number
-  agentId: number | null
-  campaignId: number
-  workRange: unknown
-  progress: unknown
-  projectId: number
-  retryCount: number
-}
-
-/**
- * Permanently fail a stale task and notify listeners. Three branches in
- * `reassignStaleTasks` reach a terminal state with the same row shape
- * (`failed` + reason + completedAt + updatedAt); centralising the body
- * keeps them from drifting.
- *
- * The UPDATE is guarded by `status IN ('assigned', 'running') AND
- * agent_id = staleTask.agentId` so a concurrent sweep (e.g. transient
- * overlap during a rolling deploy) cannot double-process the same row.
- * Returns `true` when the UPDATE actually changed a row; callers use
- * this to gate the event broadcast and aggregate refresh so a no-op
- * (row already swept by a peer) does not produce phantom transitions
- * or skew sweep metrics.
- */
-async function terminalFailStaleTask(
-  staleTask: StaleTaskRow,
-  failureReason: 'keyspace_progress_overrun' | 'max_retries_exceeded'
-): Promise<boolean> {
-  const updated = await db
-    .update(tasks)
-    .set({
-      status: 'failed',
-      failureReason,
-      completedAt: new Date(),
-      updatedAt: new Date(),
-    })
-    .where(
-      and(
-        eq(tasks.id, staleTask.taskId),
-        sql`${tasks.status} IN ('assigned', 'running')`,
-        staleTask.agentId === null ? sql`TRUE` : eq(tasks.agentId, staleTask.agentId)
-      )
-    )
-    .returning({ id: tasks.id })
-  if (updated.length === 0) {
-    return false
-  }
-  emitTaskUpdate(staleTask.projectId, staleTask.taskId, 'failed', {
-    campaignId: staleTask.campaignId,
-  })
-  await updateCampaignProgress(staleTask.campaignId)
-  return true
-}
-
-/**
- * Reassigns tasks from agents that have gone offline. Called periodically
- * by a background job (every 2 minutes via BullMQ).
- *
- * Rebalance policy when a stale task carries non-zero
- * `progress.keyspaceProgress`:
- *
- *   - If progress exceeds the task's total keyspace, the agent reported a
- *     value the implementation cannot produce - mark the task `failed`
- *     immediately (data corruption, not a retryable agent failure).
- *   - Otherwise trim `workRange.start` forward by the reported progress so
- *     the next claimant doesn't re-execute the already-cracked range.
- *     `workRange.total` is recomputed from the new start/end.
- *
- * 0% progress falls through to the existing reset-to-pending behavior
- * unchanged.
- *
- * Per-task processing is wrapped in try/catch so a transient failure on
- * one stranded task (DB serialization, broadcast error, downstream throw)
- * does not strand the rest of the batch until the next sweep tick.
- */
-export async function reassignStaleTasks(staleThresholdMs = 5 * 60 * 1000) {
-  const threshold = new Date(Date.now() - staleThresholdMs)
-
-  // Find tasks assigned to agents that haven't checked in. Carry workRange,
-  // progress, projectId, campaignId, and retryCount so the rebalance branches
-  // don't need extra queries to publish task_update events, update the
-  // campaign progress aggregate, or gate on the retry budget.
-  const staleTasks = await db
-    .select({
-      taskId: tasks.id,
-      agentId: tasks.agentId,
-      campaignId: tasks.campaignId,
-      workRange: tasks.workRange,
-      progress: tasks.progress,
-      projectId: campaigns.projectId,
-      retryCount: tasks.retryCount,
-    })
-    .from(tasks)
-    .innerJoin(agents, eq(tasks.agentId, agents.id))
-    .innerJoin(campaigns, eq(tasks.campaignId, campaigns.id))
-    .where(
-      and(
-        // Include both 'assigned' (claimed but never reported) and 'running'
-        // (claimed and reporting progress). The original 'assigned' filter
-        // missed the only state that can carry meaningful keyspaceProgress
-        // values from the agent API, leaving stranded in-progress work
-        // un-rebalanced.
-        sql`${tasks.status} IN ('assigned', 'running')`,
-        sql`${agents.lastSeenAt} < ${threshold}`,
-        // Guard against reaping a task the agent only just claimed after a
-        // stale heartbeat — without this, a slow first-heartbeat after
-        // assignment would race the sweep and yank the task back. The
-        // IS NULL branch keeps legacy/migrated rows (status='assigned' with
-        // no assigned_at) selectable instead of stranding them forever
-        // (NULL < timestamp evaluates to NULL, which the filter rejects).
-        sql`(${tasks.assignedAt} IS NULL OR ${tasks.assignedAt} < ${threshold})`
-      )
-    )
-
-  let reassigned = 0
-  let rebalanced = 0
-  let failedOverrun = 0
-  let failedMaxRetries = 0
-  let errored = 0
-  for (const staleTask of staleTasks) {
-    try {
-      const start = readWorkRangeField(staleTask.workRange, 'start')
-      const end = readWorkRangeField(staleTask.workRange, 'end')
-      const total = end > start ? end - start : 0n
-      const keyspaceProgress = readKeyspaceProgress(staleTask.progress)
-      const exceededRetries = staleTask.retryCount >= MAX_RETRIES
-
-      if (keyspaceProgress >= total && total > 0n) {
-        // Agent reported as-much-or-more work than the chunk contains. Either
-        // the agent finished the entire range but died before sending the
-        // completion message, or its report is malformed. Either way, the
-        // chunk did not flow through the normal completion path - mark
-        // failed so a fresh agent reruns the range rather than silently
-        // trusting an un-acked completion. Only count the outcome when the
-        // helper actually changed a row — a concurrent sweep that already
-        // processed this task makes the UPDATE a no-op.
-        if (await terminalFailStaleTask(staleTask, 'keyspace_progress_overrun')) {
-          failedOverrun++
-        }
-        continue
-      }
-
-      if (keyspaceProgress > 0n && keyspaceProgress < total) {
-        if (exceededRetries) {
-          // Retry budget exhausted - permanent fail. Mirrors
-          // handleTaskFailure's terminal branch; keep agentId so operators
-          // can still see which agent dropped the task.
-          if (await terminalFailStaleTask(staleTask, 'max_retries_exceeded')) {
-            failedMaxRetries++
-          }
-          continue
-        }
-
-        // Partial progress - trim workRange.start forward and re-pend.
-        const newStart = start + keyspaceProgress
-        const newTotal = end - newStart
-        // Reset reported keyspaceProgress so the next agent starts from 0
-        // within the trimmed range, but preserve auxiliary samples (speed,
-        // temperature) so the dashboard's per-task telemetry doesn't
-        // momentarily blank out across a rebalance.
-        const priorProgress =
-          staleTask.progress && typeof staleTask.progress === 'object'
-            ? (staleTask.progress as Record<string, unknown>)
-            : {}
-        const carriedProgress: Record<string, unknown> = { ...priorProgress }
-        delete carriedProgress['keyspaceProgress']
-        const rebalanceUpdated = await db
-          .update(tasks)
-          .set({
-            status: 'pending',
-            agentId: null,
-            assignedAt: null,
-            startedAt: null,
-            workRange: {
-              start: jsonSafeBigint(newStart),
-              end: jsonSafeBigint(end),
-              total: jsonSafeBigint(newTotal),
-            },
-            progress: carriedProgress,
-            retryCount: sql`${tasks.retryCount} + 1`,
-            updatedAt: new Date(),
-          })
-          .where(
-            and(
-              eq(tasks.id, staleTask.taskId),
-              sql`${tasks.status} IN ('assigned', 'running')`,
-              staleTask.agentId === null ? sql`TRUE` : eq(tasks.agentId, staleTask.agentId)
-            )
-          )
-          .returning({ id: tasks.id })
-        if (rebalanceUpdated.length === 0) {
-          continue
-        }
-        emitTaskUpdate(staleTask.projectId, staleTask.taskId, 'pending', {
-          campaignId: staleTask.campaignId,
-        })
-        await updateCampaignProgress(staleTask.campaignId)
-        rebalanced++
-        continue
-      }
-
-      if (exceededRetries) {
-        // 0% / unreadable progress but retry budget exhausted - permanent fail.
-        if (await terminalFailStaleTask(staleTask, 'max_retries_exceeded')) {
-          failedMaxRetries++
-        }
-        continue
-      }
-
-      // 0% progress or unreadable range - reset to pending unchanged.
-      const resetUpdated = await db
-        .update(tasks)
-        .set({
-          status: 'pending',
-          agentId: null,
-          assignedAt: null,
-          startedAt: null,
-          retryCount: sql`${tasks.retryCount} + 1`,
-          updatedAt: new Date(),
-        })
-        .where(
-          and(
-            eq(tasks.id, staleTask.taskId),
-            sql`${tasks.status} IN ('assigned', 'running')`,
-            staleTask.agentId === null ? sql`TRUE` : eq(tasks.agentId, staleTask.agentId)
-          )
-        )
-        .returning({ id: tasks.id })
-      if (resetUpdated.length === 0) {
-        continue
-      }
-      emitTaskUpdate(staleTask.projectId, staleTask.taskId, 'pending', {
-        campaignId: staleTask.campaignId,
-      })
-      await updateCampaignProgress(staleTask.campaignId)
-      reassigned++
-    } catch (err) {
-      errored += 1
-      logger.error(
-        {
-          err,
-          taskId: staleTask.taskId,
-          agentId: staleTask.agentId,
-          campaignId: staleTask.campaignId,
-          projectId: staleTask.projectId,
-        },
-        'reassignStaleTasks: per-task processing threw — sibling stale tasks continue'
-      )
-    }
-  }
-
-  return { reassigned, rebalanced, failedOverrun, failedMaxRetries, errored }
-}
-
 // ─── Task Queries ───────────────────────────────────────────────────
 
 /**
@@ -1284,3 +873,11 @@ export async function listTasksByAgent(
 
   return projectAgentTaskRows(rows)
 }
+
+// ─── Re-exports from ./tasks/* submodules (CQ-H1 split) ───────────
+//
+// The retry/failure machinery (handleTaskFailure, reassignStaleTasks,
+// MAX_RETRIES) lives in its own file so the parent service stays under
+// the 800-line budget. Re-exporting here keeps caller import paths
+// stable (services/tasks.js continues to surface the same symbols).
+export { handleTaskFailure, MAX_RETRIES, reassignStaleTasks } from './tasks/retry.js'
